requirements.txt
discord.py>=2.7.1,<3.0
yt-dlp>=2026.3.17
python-dotenv>=1.2.2
pydub~=0.25.1
numpy~=1.23.1
g2p_en
